Into - Update 1
Got some basic placeholders, and am working up the screenfade code. It
should be handy when we impliment the claustrophobia game mechanic, as
well as for cutscenes. I learned that cashing multiple audio sources on
a single object doesn't work so I'll have to split them into multiple
objects. So far so good. Please don't mess with it anyone ok? Should be
finished up inside the week. Gotta stop and go recruiting.